fix: 폴더 이름 수정 & Theme 페이지 생성
