+++ frontend/src/components/Common/CardItem/CardList.styles.ts
