+++ frontend/src/components/GiftThema/Thema/ThemeItem.tsx
@@ -1,4 +1,4 @@
-import { ThemeCard } from '@/components/GiftThema/Thema/ThemeList.styles.ts';
+import { ThemeCard } from '@/components/GiftThema/Thema/ThemeItem.styles.ts';
 
 interface Props {
   image: string;
@@ -6,7 +6,7 @@ interface Props {
   onClick: (event: React.MouseEvent<HTMLDivElement>) => void;
 }
 
-export default function ThemeList({ image, name, onClick }: Props) {
+export default function ThemeItem({ image, name, onClick }: Props) {
   return (
     <ThemeCard onClick={onClick}>
       <img src={image} alt={name} />
