+++ frontend/src/components/GiftThema/ThemaItem/ThemeInfo.tsx
@@ -1,34 +1,17 @@
-import { useNavigate, useParams } from 'react-router-dom';
-import useFetchThemesInfo from '@/hooks/fetch/useFetchThemesInfo.ts';
-import Header from '@/components/Header/Header.tsx';
 import {
   ThemeInfoHeader,
   ThemeInfoWrapper,
 } from '@/components/GiftThema/ThemaItem/ThemeInfo.styles.ts';
-import ThemeProducts from '@/components/GiftThema/ThemaItem/ThemeProducts.tsx';
-import { useEffect } from 'react';
 import {
   ProductsError,
   ProductsLoading,
 } from '@/components/GiftThema/ThemaItem/ThemeProducts.styles.ts';
-import { PATH } from '@/constants/path.ts';
 
-export default function ThemeInfo() {
-  const navigate = useNavigate();
-  const { id } = useParams<{ id: string }>();
-  const themeId = Number(id);
-  const { themeInfo, loading, error, statusCode } = useFetchThemesInfo(themeId);
+export default function ThemeInfo({ loading, error, themeInfo }) {
 
-  useEffect(() => {
-    if (statusCode === 404) {
-      navigate(`${PATH.HOME}`);
-    }
-  }, [statusCode, navigate]);
 
   return (
     <ThemeInfoWrapper>
-      <Header />
-
       {loading ? (
         <ProductsLoading>로딩 중...</ProductsLoading>
       ) : error || !themeInfo ? (
@@ -40,8 +23,6 @@ export default function ThemeInfo() {
           <p>{themeInfo.description}</p>
         </ThemeInfoHeader>
       )}
-
-      <ThemeProducts themeId={themeId} />
     </ThemeInfoWrapper>
   );
 }
